Small typing fix and more robust model handling
Changed l <= lam[spe] to l < lam[spe] to fit more in line with the original implementation.

Also noticed, that not all models train properly, sometimes leaving empty projectors which means the model can not predict these elements even though they were in the training data. I have not fixed this yet, but a warning is displayed if this is the case!