Refactored big parts of descriptor calculation
To reduce the number of matrix refits and flattens, i changed how the descriptors were calculated in the equicomb function. This lead to major changes in the code that used the descriptors. I also added a way to perform the dot product by using two flattened matrices.

This all for a speedup of about ~0.5s (make of that what you will)