Added better ID system for TSCB files
IDs are based on the atomic number and the fractional coordinates of each atom in the cif file. Currently only implemented for SALTED calculations, to cahnge this at the bottom in every scattering calculation you would need to replace the labels given to tsc_block with the new labels. Also reading is not implemented on the OLex2 side, so don't....